added error handelr and componentized prompt
